+++ src/data/consume/drakeheart-mutagen.ts
@@ -1,0 +1,106 @@
+import { AssistantAction } from "action.ts";
+import { EffectSource } from "foundry-pf2e";
+import { AssistantMessage } from "message.ts";
+
+export const actions: AssistantAction[] = [
+    {
+        trigger: "consume",
+        predicate: ["consumable:drakeheart-mutagen-lesser"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.GBBjw61g4ekJymT0", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:drakeheart-mutagen-moderate"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.BV8RPntjc9FUzD3g", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:drakeheart-mutagen-greater"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.qwoLV4awdezlEJ60", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:drakeheart-mutagen-major"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-srd.Item.M4ZOHOlne43ArjOC", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+];